Optimize file descriptors
The uint32_t type is much better for this purpose.

+++ include/kernel/task.h
@@ -74,8 +74,8 @@ struct task {
 		uint32_t state;
 		void (*release)(struct task *task);
 		void (*clone)(struct task *task, struct task *new_task);
-		uint64_t *table;
-		uint64_t _table[TASK_FD_STATIC_COUNT];
+		uint32_t *table;
+		uint32_t _table[TASK_FD_STATIC_COUNT];
 	} fd;
 };
